refactor/modules: move iterators into a separate module
Makes it easier to maintain the codebase.
